Getting things back together including a patch from Ajay Lulia
git-svn-id: http://joomlacode.org/svn/joomla/development/branches/update@11946 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ libraries/joomla/installer/adapters/library.php
@@ -3,43 +3,25 @@
  * @version		$Id:plugin.php 6961 2007-03-15 16:06:53Z tcp $
  * @package		JLibMan
  * @subpackage	Installer
- * @copyright 	Copyright (C) 2008 Open Source Matters
- * @license		GNU/GPL, see LICENSE.php
- * Joomla! is free software. This version may have been modified pursuant
- * to the GNU General Public License, and as distributed it includes or
- * is derivative of works licensed under the GNU General Public License or
- * other free or open source software licenses.
- * See COPYRIGHT.php for copyright notices and details.
+ * @copyright 	Copyright (C) 2005 - 2009 Open Source Matters, Inc. All rights reserved.
+ * @license		GNU General Public License <http://www.gnu.org/copyleft/gpl.html>
  */
 
 // Check to ensure this file is within the rest of the framework
-defined('JPATH_BASE') or die();
+defined('JPATH_BASE') or die;
 
-jimport( 'joomla.installer.librarymanifest' );
+jimport('joomla.installer.librarymanifest');
 jimport('joomla.base.adapterinstance');
 
 /**
  * Library installer
- * 
- * @author 		Sam Moffatt <pasamio@gmail.com>
+ *
  * @package		Joomla.Framework
  * @subpackage	Installer
  * @since		1.6
  */
 class JInstallerLibrary extends JAdapterInstance
 {
-	/**
-	 * Constructor
-	 *
-	 * @access	protected
-	 * @param	object	$parent	Parent object [JInstaller instance]
-	 * @return	void
-	 * @since	1.5
-	 */
-	function __construct(&$parent)
-	{
-		$this->parent =& $parent;
-	}
 
 	/**
 	 * Custom install method
@@ -51,8 +33,8 @@ class JInstallerLibrary extends JAdapterInstance
 	function install()
 	{
 		// Get the extension manifest object
-		$manifest =& $this->parent->getManifest();
-		$this->manifest =& $manifest->document;
+		$manifest = &$this->parent->getManifest();
+		$this->manifest = &$manifest->document;
 
 		/**
 		 * ---------------------------------------------------------------------------------------------
@@ -61,17 +43,17 @@ class JInstallerLibrary extends JAdapterInstance
 		 */
 
 		// Set the extensions name
-		$name =& $this->manifest->getElementByPath('name');
+		$name = &$this->manifest->getElementByPath('name');
 		$name = JFilterInput::clean($name->data(), 'string');
 		$element = str_replace('.xml','',basename($this->parent->getPath('manifest')));
 		$this->set('name', $name);
 		$this->set('element', $element);
 
-		$db =& $this->parent->getDBO();
+		$db = &$this->parent->getDbo();
 		$db->setQuery('SELECT extension_id FROM #__extensions WHERE type="library" AND element = "'. $element .'"');
 		$result = $db->loadResult();
-		if($result) { // already installed, can we upgrade?
-			if($this->parent->getOverwrite() || $this->parent->getUpgrade()) { 
+		if ($result) { // already installed, can we upgrade?
+			if ($this->parent->getOverwrite() || $this->parent->getUpgrade()) {
 				// we can upgrade, so uninstall the old one
 				$installer = new JInstaller(); // we don't want to compromise this instance!
 				$installer->uninstall('library', $result);
@@ -88,11 +70,11 @@ class JInstallerLibrary extends JAdapterInstance
 		if (is_a($description, 'JSimpleXMLElement')) {
 			$this->parent->set('message', $description->data());
 		} else {
-			$this->parent->set('message', '' );
+			$this->parent->set('message', '');
 		}
 
 		// Set the installation path
-		$element =& $this->manifest->getElementByPath('files');
+		$element = &$this->manifest->getElementByPath('files');
 		$group = $this->manifest->getElementByPath('libraryname');
 		$group = $group->data();
 		if (!empty($group)) {
@@ -132,7 +114,7 @@ class JInstallerLibrary extends JAdapterInstance
 			$this->parent->abort();
 			return false;
 		}
-		
+
 		/**
 		 * ---------------------------------------------------------------------------------------------
 		 * Extension Registration
@@ -147,9 +129,9 @@ class JInstallerLibrary extends JAdapterInstance
 		$row->protected = 0;
 		$row->access = 0;
 		$row->client_id = 0;
-		$row->data = ''; // custom data
 		$row->params = $this->parent->getParams();
-		$row->manifestcache = $this->parent->generateManifestCache();
+		$row->data = ''; // custom data
+		$row->manifest_cache = $this->parent->generateManifestCache();
 		if (!$row->store()) {
 			// Install failed, roll back changes
 			$this->parent->abort(JText::_('Library').' '.JText::_('Install').': '.$db->stderr(true));
@@ -183,8 +165,8 @@ class JInstallerLibrary extends JAdapterInstance
 	function update() {
 		// since this is just files, an update removes old files
 		// Get the extension manifest object
-		$manifest =& $this->parent->getManifest();
-		$this->manifest =& $manifest->document;
+		$manifest = &$this->parent->getManifest();
+		$this->manifest = &$manifest->document;
 
 		/**
 		 * ---------------------------------------------------------------------------------------------
@@ -193,21 +175,22 @@ class JInstallerLibrary extends JAdapterInstance
 		 */
 
 		// Set the extensions name
-		$name =& $this->manifest->getElementByPath('name');
+		$name = &$this->manifest->getElementByPath('name');
 		$name = JFilterInput::clean($name->data(), 'string');
 		$element = str_replace('.xml','',basename($this->parent->getPath('manifest')));
 		$this->set('name', $name);
 		$this->set('element', $element);
 		$installer = new JInstaller(); // we don't want to compromise this instance!
-		$db->setQuery('SELECT extensionid FROM #__extensions WHERE type="library" AND element = "'. $element .'"');
+		$db = &$this->parent->getDbo();
+		$db->setQuery('SELECT extension_id FROM #__extensions WHERE type="library" AND element = "'. $element .'"');
 		$result = $db->loadResult();
-		if($result) { // already installed, which would make sense
+		if ($result) { // already installed, which would make sense
 			$installer->uninstall('library', $result);
 		}
 		// now create the new files
 		return $this->install();
 	}
-	
+
 	/**
 	 * Custom uninstall method
 	 *
@@ -220,23 +203,23 @@ class JInstallerLibrary extends JAdapterInstance
 	{
 		// Initialize variables
 		$retval = true;
-		
+
 		// First order of business will be to load the module object table from the database.
 		// This should give us the necessary information to proceed.
 		$row = & JTable::getInstance('extension');
-		if ( !$row->load((int) $id) || !strlen($row->element)) {
+		if (!$row->load((int) $id) || !strlen($row->element)) {
 			JError::raiseWarning(100, JText::_('ERRORUNKOWNEXTENSION'));
 			return false;
 		}
-		
+
 		// Is the library we are trying to uninstall a core one?
 		// Because that is not a good idea...
 		if ($row->protected) {
 			JError::raiseWarning(100, JText::_('Library').' '.JText::_('Uninstall').': '.JText::sprintf('WARNCOREMODULE', $row->name)."<br />".JText::_('WARNCOREMODULE2'));
 			return false;
 		}
-		
-		$manifestFile = JPATH_MANIFESTS.DS.'libraries' . DS . $row->element .'.xml'; 
+
+		$manifestFile = JPATH_MANIFESTS.DS.'libraries' . DS . $row->element .'.xml';
 
 		// Because libraries may not have their own folders we cannot use the standard method of finding an installation manifest
 		if (file_exists($manifestFile))
@@ -244,8 +227,8 @@ class JInstallerLibrary extends JAdapterInstance
 			$manifest = new JLibraryManifest($manifestFile);
 			// Set the plugin root path
 			$this->parent->setPath('extension_root', JPATH_ROOT.DS.'libraries'.DS.$manifest->libraryname);
-			
-			$xml =& JFactory::getXMLParser('Simple');
+
+			$xml = &JFactory::getXMLParser('Simple');
 
 			// If we cannot load the xml file return null
 			if (!$xml->loadFile($manifestFile)) {
@@ -258,7 +241,7 @@ class JInstallerLibrary extends JAdapterInstance
 			 * @todo: Remove backwards compatability in a future version
 			 * Should be 'extension', but for backward compatability we will accept 'install'.
 			 */
-			$root =& $xml->document;
+			$root = &$xml->document;
 			if ($root->name() != 'install' && $root->name() != 'extension') {
 				JError::raiseWarning(100, JText::_('Library').' '.JText::_('Uninstall').': '.JText::_('Invalid manifest file'));
 				return false;
@@ -266,10 +249,10 @@ class JInstallerLibrary extends JAdapterInstance
 
 			$this->parent->removeFiles($root->getElementByPath('files'), -1);
 			JFile::delete($manifestFile);
+
 		} else {
-			$row->delete($row->extension_id);
 			// remove this row entry since its invalid
-			$row->delete($row->extensionid);
+			$row->delete($row->extension_id);
 			unset($row);
 			JError::raiseWarning(100, 'Library Uninstall: Manifest File invalid or not found');
 			return false;
@@ -277,24 +260,24 @@ class JInstallerLibrary extends JAdapterInstance
 
 		// TODO: Change this so it walked up the path backwards so we clobber multiple empties
 		// If the folder is empty, let's delete it
-		if(JFolder::exists($this->parent->getPath('extension_root'))) {
-			if(is_dir($this->parent->getPath('extension_root'))) {
+		if (JFolder::exists($this->parent->getPath('extension_root'))) {
+			if (is_dir($this->parent->getPath('extension_root'))) {
 				$files = JFolder::files($this->parent->getPath('extension_root'));
+				if (!count($files)) {
 					JFolder::delete($this->parent->getPath('extension_root'));
-		}
 				}
-		$row->delete($row->extension_id);
+			}
 		}
-		
-		$row->delete($row->extensionid);
+
+		$row->delete($row->extension_id);
 		unset($row);
 
 		return $retval;
 	}
-	
+
 /**
 	 * Custom discover method
-	 * 
+	 *
 	 * @access public
 	 * @return array(JExtension) list of extensions available
 	 * @since 1.6
@@ -304,20 +287,20 @@ class JInstallerLibrary extends JAdapterInstance
 		$file_list = JFolder::files(JPATH_MANIFESTS.DS.'libraries','\.xml$');
 		foreach($file_list as $file) {
 			$file = JFile::stripExt($file);
-			$extension =& JTable::getInstance('extension');
-			$extension->type = 'library';
-			$extension->client_id = 0;
-			$extension->element = $file;
-			$extension->name = $file;
-			$extension->state = -1;
+			$extension = &JTable::getInstance('extension');
+			$extension->set('type', 'library');
+			$extension->set('client_id', 0);
+			$extension->set('element', $file);
+			$extension->set('name', $file);
+			$extension->set('state', -1);
 			$results[] = $extension;
 		}
 		return $results;
 	}
-	
+
 	/**
 	 * Custom discover_install method
-	 * 
+	 *
 	 * @access public
 	 * @param int $id The id of the extension to install (from #__discoveredextensions)
 	 * @return void
@@ -328,28 +311,29 @@ class JInstallerLibrary extends JAdapterInstance
 		 * There are two parts to a library which are disjunct in their locations
 		 * 1) The manifest file (stored in /JPATH_MANIFESTS/libraries)
 		 * 2) The actual files (stored in /JPATH_LIBRARIES/libraryname)
-		 * Thus installation of a library is the process of dumping files 
+		 * Thus installation of a library is the process of dumping files
 		 * in two different places. As such it is impossible to perform
 		 * any operation beyond mere registration of a library under the presumption
+		 * that the files exist in the appropriate location so that come uninstall
+		 * time they can be adequately removed.
 		 */
+
 		$manifestPath = JPATH_MANIFESTS . DS . 'libraries' . DS . $this->parent->extension->element . '.xml';
 		$this->parent->manifest = $this->parent->isManifest($manifestPath);
-		$manifestPath = JPATH_MANIFESTS . DS . 'libraries' . DS . $this->parent->_extension->element . '.xml';
+		$this->parent->setPath('manifest', $manifestPath);
+		$manifest_details = JApplicationHelper::parseXMLInstallFile($this->parent->getPath('manifest'));
 		$this->parent->extension->manifest_cache = serialize($manifest_details);
 		$this->parent->extension->state = 0;
 		$this->parent->extension->name = $manifest_details['name'];
 		$this->parent->extension->enabled = 1;
 		$this->parent->extension->params = $this->parent->getParams();
-		if($this->parent->extension->store()) {
-		$this->parent->_extension->enabled = 1;
-		$this->parent->_extension->params = $this->parent->getParams();
-		if($this->parent->_extension->store()) {
+		if ($this->parent->extension->store()) {
 			return true;
 		} else {
 			JError::raiseWarning(101, JText::_('Plugin').' '.JText::_('Discover Install').': '.JText::_('Failed to store extension details'));
 			return false;
 		}
-	}		
+	}
 
 }
